add new options for rs menu and fix gapps install

+++ romswitcher/system_format.sh
@@ -2,12 +2,15 @@
 
 BB="busybox"
 
-$BB umount -f /system
-$BB mkdir -p /system
-system=/.firstrom/media/.secondrom/system.img
-$BB mount -t ext4 -o rw $system /system
-$BB rm -rf /system/*
-$BB rm -rf /system/.*
-$BB mke2fs -F -T ext4 $system || exit 1
+if [ "$1" == "secondary" ] ; then
+   system=/.firstrom/media/.secondrom/system.img
+   $BB umount -f /system
+   $BB mkdir -p /system
+
+   $BB mount -t ext4 -o rw $system /system || exit 1
+   $BB rm -rf /system/*
+   $BB rm -rf /system/.*
+   $BB mke2fs -F -T ext4 $system || exit 1
+fi
 
 exit 0
